requirement.txt
anyio==3.6.2
click==8.1.3
fastapi==0.90.1
h11==0.14.0
idna==3.4
Library==0.0.0
psycopg2==2.9.5
pydantic==1.10.4
PyJWT==2.6.0
PyMySQL==1.0.2
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.23.1
typing_extensions==4.4.0
uvicorn==0.20.0
